requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pillow
numpy
torch
torchvision
ultralytics
opencv-python-headless
easyocr
